project_bank/requirements.txt
pandas
sqlalchemy
great_expectations
prefect
psycopg2-binary
alembic 
FastAPI
numpy
sqlalchemy.orm
sqlalchemy.sql
Datetime
matplotlib.pyplot
Seaborn
Scipy.stats
Logging
great_expectations
Pathlib 
Pytest
os